+++ ubo_app/utils/apt.py
@@ -8,27 +8,30 @@ import subprocess
 from ubo_app.logging import logger
 
 
-def install_package(package_name: str) -> None:
+def install_package(packages: str | list[str], /, *, force: bool = False) -> None:
     """Install a package using the APT package manager."""
     import apt  # pyright: ignore [reportMissingModuleSource]
 
+    package_names = packages if isinstance(packages, list) else [packages]
+
     cache = apt.Cache()
-    logger.debug('Installing package...', extra={'package': package_name})
-    if package_name in cache:
-        pkg = cache[package_name]
-        if not pkg.is_installed:
-            pkg.mark_install()
-            cache.commit()
-            logger.info(
-                'Package installed successfully.',
-                extra={'package': package_name},
-            )
+    logger.debug('Installing packages...', extra={'packages': package_names})
+    for package_name in package_names:
+        if package_name in cache:
+            pkg = cache[package_name]
+            if not pkg.is_installed or force:
+                pkg.mark_install()
+                cache.commit()
+                logger.info(
+                    'Package installed successfully.',
+                    extra={'package': package_name},
+                )
+            else:
+                msg = f'Package {package_name} is already installed.'
+                raise ValueError(msg)
         else:
-            msg = f'Package {package_name} is already installed.'
+            msg = f'Package {package_name} not found in APT cache.'
             raise ValueError(msg)
-    else:
-        msg = f'Package {package_name} not found in APT cache.'
-        raise ValueError(msg)
 
 
 def uninstall_package(package_name: str) -> None: